+++ src/main/java/cart/controller/advice/AdminControllerAdvice.java
@@ -1,5 +1,7 @@
-package cart.controller;
+package cart.controller.advice;
 
+import cart.controller.AdminController;
+import cart.controller.AdminProductController;
 import cart.dto.ErrorResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,8 +18,6 @@ import static org.springframework.http.HttpStatus.NOT_FOUND;
 @RestControllerAdvice(assignableTypes = {AdminController.class, AdminProductController.class})
 public class AdminControllerAdvice {
 
-    private static final String UNEXPECTED_ERROR_MESSAGE = "네트워크 연결 상태가 좋지 않습니다 잠시 후 다시 시도해 주세요.";
-
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
@@ -29,13 +29,6 @@ public class AdminControllerAdvice {
         return ResponseEntity.badRequest().body(new ErrorResponse(errorMessage));
     }
 
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> handleUnexpectedException(final Exception exception) {
-        final String errorMessage = exception.getMessage();
-        logger.error("error : {}", errorMessage);
-
-        return ResponseEntity.internalServerError().body(new ErrorResponse(UNEXPECTED_ERROR_MESSAGE));
-    }
 
     @ExceptionHandler(NoSuchElementException.class)
     public ResponseEntity<ErrorResponse> handleNoSuchElementException(final NoSuchElementException exception) {
